refactor(app): align function store module naming
