+++ crates/coral-app/src/functions/mod.rs
@@ -1,11 +1,11 @@
 //! Function lifecycle and inventory workflow.
 #![expect(
     dead_code,
-    reason = "function inventory and artifact storage are introduced before runtime and service callers in the split app stack"
+    reason = "function inventory and artifact stores are introduced before runtime and service callers in the split app stack"
 )]
 
 pub(crate) mod manager;
 pub(crate) mod model;
-mod storage;
+mod store;
 
 pub(crate) use model::FunctionName;
