+++ crates/coral-app/src/functions/store.rs
